ErrorHandler_Message_t.py contains tests that publish messages to test the erorhandler only. It assumes the ErrorHandler component is running (no other components are running). This test is part of a suite.
ReqInj_Message_t.py contains message publications (needs to be formatted
using the testcase class) and assumes that the ReqInj component, JobCreator,
JobSubmitter, ErrorHandler is running.  The messages in this file are for
the ReqInj component which acts on them and generates messages for the
JobCreator which generates messages for the JobSubmitter, etc..
The payload of the messages are referenses to workflow spec files
(these files are not part of CVS!).


--------------------------------------
Frank van Lingen
California Institute of Technology
CA 91125 Pasadena
United States
--------------------------------------
Mail Code:356-48
bld      :340 Lauritsen
email    :fvlingen@caltech.edu
im(aim)  :marcellus0872
tel      :(+1) 626 395 3862
mobile   :(+1) 310 968 5584
url      :http://www.van-lingen.name/
--------------------------------------