Added use of the BOSS '-c' option (allowing specify directory containing the configuration files). dbConfig['configDir'] value is for the moment hardcoded to the value: $PRODAGENT_WORKDIR/bosscfg.
Added registration of the program type CMSSW (will be distributed in
the next BOSS tag: for the moment, if the schema file is not found, the
registration is skipped).

Used a different name for the Real Time Database, just appending '_rt"
to the BOSS db name: a better solution has to be found.

Added the sql schema name detection from the BOSS stdout, to avoid
problems with old ones for some reason not removed.