Extended the ProdMgr interface module, and moved service management methods to a management module.
ProdAgent modules can interact with a prodmanager by importing
the api:

import ProdMgrInterface.Interface as ProdMgrAPI

It shields the ProdAgent modules from any interaction sessions and makes
it easy to make changes if necessary.

Interaction are logged by default and need to be commited by the module
making the call. e.g.:

allocations=ProdMgrAPI.acquireAllocation(self.serverURL,self.request_id,15)
...do some other things....
ProdMgrAPI.commit()

If the call can not be commited, and the prodagent crashes it is possible
to do a retrieve/resend request to the server where the prodmgr is being hosted

This approach ensures the prodmgr database is kept consistent while also preventing
possible loss of information on the prodagent side.

Frank.