Added functionality to component for job retrieval.
Large part of the code deals with robust interaction with the
ProdMgr to prevent information loss when crashes occur.

Todo:
-The priorityQueue class(es) need to be implemented in a database backend
-Add a table in the prodagent database that tracks the state of the ProdMgrInterface
component in a persistent way. This table can also be used by other components
(should that be necessary) to track their state
-A global cursor so we can do an effective rollback if something goes
wrong without loss of information or inconsistencies in the database. This
is also something that will be needed in the other components. Note:
part of the database interaction in the ProdMgrInterface should be immediate
commited as it deals with interaction with the ProdMgr (global schema), other
parts that deal with the local schema should generally not.
-Add interaction with the prodmgr for job success and job failure.

Note:
- The current code contains a lot of debug statements and comments to
describe the process.

Frank.