Added extensions for file coupling. After a job success the prodmgr interface registers generated files. The merge sensor than calls the File.merge method once files are registered which will emit a "ProdMgrInterface:JobSuccess" once all files of a job are safely merged. This event is caught and processed and the prodmgr is contacted if necessary.
In the config file there is an override option for this for the prodmgr interface
which can also call the File.merge method for testing purposes.

The new aux files are used for handling jobs that succeed in the JobSuccessFinal
state and jobs that are failed and did not generate any files in the JobSuccess
state (as the merge sensor will not give any feedback on the latter)

Frank.