-Removed some inconsistent behaviour seen during testing -Finalized the WorkflowEntities.Job interface -Inserted a mapping in the JobState interface to the Job interface. Once all components use the new JobState interface (which is identical to the old one (including the underlying tables), we can activate the new interface by turning on the wrapper after which the new workflow entities Job object will be transparently used through the old interface.
Once we use the new interface we can change the foreign key contraints of the trigger
table and remove the old jobstate tables.

Methods that do not have the wrapper (as they are not used in the code) from the old
interface are: jobReports,retrieveJobIds (if necessary we can extend the new interface
with these methods but not migrate them.