Added two methods in the ProdAgent/WorkflowEntities/Aux.py that splits a string or gets the separator. These methods are there to migrate from the '_' splitter to the '@' spliter. If you split the job id to acquire the asociated workflow without  querying the databasee replace the split with Aux.split( )
The module contains an old separator and a new one and differentiates between
the new and old onw depending on the structure of the input string.

These methods will enable us to migrate to the new separator without any disruption
of the services.

Note: these methods are not needed if you do splits on other strings than the job id.


def split(inputString):

def getSeparator(inputString):