Changed state assignment for jobs:
-state changes are not checked anymore
-it is possible (before run time) to add your own states to
the we_Job.status column and thereby augment the default states.

The current implementation expects that the we_Job.enum type in the
prodagent sql statement is changed to deal with additional state information.

This was favorered instead of foreign key to another table for two reasons:

-updating a state with a foreign key constraint is more expensive (takes
approximately 1.5 to 2 time as long.
-Although states can be different in different applications (CRAB, PRODAGENT,
etc...) we do not expect to have 10s or 100s of states (which means we can
use a simple enum.
-States can be dynamic per application, but within the application they are
static, which means they can be defined at creation time of the database.


But beware:

There are 8 'base' states: 'register','released','create','submit',
'inProgress','finished','reallyFinished','failed' which can be extended
by defining additional states. Some of these base states are used to update
retry information. Make sure that when you change the states you do not
remove the ones that update retrye information.

Retry information is for example updated in the 'submit' state thus if you decide
to remove the default states retry updates can not be guaranteed.

Same goes for registering a failure:

It is possible to define additional states, however
retry information is only updated if there is a failure of type:
'create','submit', or 'run'