Tweaks needed to add Express,Harvesting, and Skim job types so that data ops has more knobs to play with.
The change in RunResTools changes an or specific to "Processing" and "Repack"
job types to a != "Merge" opening the possibility for any kind of job type.

The DQMInjector tweak actually switches it over to the "Harvesting" job type

To do here is possibly rigging the workflow creation tools to make "Skim"
type workflows if so desired.